+++ pkgs/ffigen/lib/ffigen.dart
@@ -55,3 +55,5 @@ export 'src/config_provider.dart'
         macSdkUri,
         xcodePath,
         xcodeUri;
+export 'src/config_provider/public_ast.dart';
+export 'src/config_provider/public_visitor.dart';