+++ pkgs/ffigen/lib/src/code_generator/objc_methods.dart
@@ -324,6 +324,8 @@ class ObjCMethod extends AstNode with HasLocalScope {
   bool get isProperty =>
       kind == ObjCMethodKind.propertyGetter ||
       kind == ObjCMethodKind.propertySetter;
+  bool get isPropertyGetter => kind == ObjCMethodKind.propertyGetter;
+  bool get isPropertySetter => kind == ObjCMethodKind.propertySetter;
   bool get isRequired => !isOptional;
   bool get isInstanceMethod => !isClassMethod;
   bool get unavailable => apiAvailability.availability == Availability.none;
@@ -509,7 +511,7 @@ class ObjCMethod extends AstNode with HasLocalScope {
     // Evaluate targetStr and msgSendParams first to populate localVars.
     late String targetStr;
     if (isClassMethod) {
-      targetStr = (target as ObjCInterface).classObject.name;
+      targetStr = (target as ObjCInterface).classObject!.name;
     } else {
       targetStr = target.convertDartTypeToFfiDartType(
         context,