+++ pkgs/ffigen/test/header_parser_tests/record_use_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:ffigen/src/config_provider.dart';
+import 'package:ffigen/ffigen.dart';
 import 'package:ffigen/src/header_parser.dart' show parse;
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
@@ -20,9 +20,16 @@ void main() {
         functions: Functions(
           include: (decl) => true,
           recordUse: (decl) => true,
-          rename: (decl) =>
-              decl.originalName == 'sum' ? 'add' : decl.originalName,
         ),
+        visitors: [
+          Visitor(
+            visitFunc: (node) {
+              if (node.name == 'sum') {
+                node.name = 'add';
+              }
+            },
+          ),
+        ],
         output: Output(
           dartFile: Uri.file('unused.dart'),
           style: const NativeExternalBindings(),